Add files via upload
The leaf disease detection system is built using python and its modules. This project is based on image processing and is used to detect the disease of the plants by taking and uploading pictures of the leaf that is affected by any kind of fungal infection like yellow fungus , black fungus,molds, blight,etc.